Add files via upload
This is the second iteration of the ExtendedConnectX project. In this version, we have two implementation of the game board, the regular way and using a HashMap. We have also added extra functionality by adding in the ability to change the size of the board, change the number to win, and change how many players can play (up to 10 players) and even allow them to pick the character they want to represent their token!