Add files via upload
This includes test cases to test that my program actually works and doesn't break any of my pre/post conditions or invariants. To do this on IntelliJ, simply move the test files in Test/cpsc2150/extendedConnectX and run them. In the terminal, simply move to the src and run "make test" and to implement the GameBoard.java tests, run "testGB" and for GameBoardMem.java, run "testGBMem"